Get rid of CBaseMerkleTx.
Since CAuxPow is now decoupled from CMerkleTx, we do not need the
CBaseMerkleTx class anymore.  In this commit, we move the Merkle tx
stuff back to wallet/wallet.h, where it is upstream.  This reduces
diffs to Bitcoin.
